+++ CHANGELOG.md
@@ -6,6 +6,12 @@ will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Token generation, storing and copying to server
+
 ## [0.7.1] - 2020/10/04
 
 - Update the change log